Revert "Update MSM_Limiter To 3.6 And Frequencies Fixed"
This reverts commit 3067135c85a86f649206d961d5833a7b6f841b52.
